pyproject.toml
[project]
name = "blogdjango"
version = "1.0.0"
description = "Blog application from Antonio Melé's book \"Django 4 By Example\""
authors = [
    {name = "Aleksan.Jhangiryan",email = "lernagorc90@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "django (==4.1.0)",
    "sqlparse (==0.4.2)",
    "asgiref (==3.5.2)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-taggit (==3.0.0)",
    "markdown (==3.4.1)",
    "psycopg2-binary (==2.9.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
skip = ["migrations", ".venv", "venv"]
